Add files via upload
Amma Pickles and Snacks is a website that brings the taste of home to everyone. It offers a variety of homemade veg and non-veg pickles, and traditional snacks. The site is simple to use browse, add to cart, and order your favorites. Made using HTML, CSS, JavaScript, and Flask. Hosted on AWS to make it easily accessible anytime.